feat: add local dev volumes
Add volumes to function containers for storage service emulation, etc.
